refactor: isolation data dependency of ClassControllerClassController
