+++ src/components/StylePreview/index.tsx
@@ -1,7 +1,7 @@
 import { StylePreviewerProvider } from "./context";
 import {
   HighlightedStyleElement,
-  ClassSelectList,
+  ClassController,
   ModeToggleController,
 } from "./ui";
 import type { StylePreviewerProps, ClassInfo } from "./type";
@@ -104,7 +104,92 @@ const StylePreviewer = <T extends Record<string, ClassInfo>>({
               )}
             </ModeToggleController>
           </div>
-          <ClassSelectList />
+          <ClassController>
+            {(
+              mode,
+              classInfo,
+              selectedClass,
+              updateClassState,
+              resetClassState
+            ) => (
+              <div>
+                {Object.entries(classInfo).map(([key, value]) => {
+                  const handleMouseLeave = () => {
+                    if (mode === "click") return;
+                    resetClassState();
+                  };
+
+                  return (
+                    <div key={key}>
+                      <h3 style={{ marginBottom: "10px" }}>{key}</h3>
+                      <ClassItem
+                        name={key}
+                        isSelected={
+                          selectedClass?.elementKey === key &&
+                          selectedClass?.type === "className"
+                        }
+                        onMouseEnter={() => {
+                          mode === "hover" &&
+                            updateClassState({
+                              elementKey: key,
+                              type: "className",
+                            });
+                        }}
+                        onMouseLeave={handleMouseLeave}
+                        onClick={() =>
+                          mode === "click" &&
+                          updateClassState({
+                            elementKey: key,
+                            type: "className",
+                          })
+                        }
+                      />
+                      {value.classNames && (
+                        <div
+                          style={{
+                            borderTop: "1px solid #eee",
+                            marginTop: "10px",
+                            paddingTop: "10px",
+                          }}
+                        >
+                          {Object.entries(value.classNames).map(
+                            ([propertyName, _]) => (
+                              <ClassItem
+                                key={propertyName}
+                                name={`classNames.${propertyName}`}
+                                isSelected={
+                                  selectedClass?.elementKey === key &&
+                                  selectedClass?.type === "classNames" &&
+                                  selectedClass?.propertyName === propertyName
+                                }
+                                onMouseEnter={() => {
+                                  mode === "hover" &&
+                                    updateClassState({
+                                      elementKey: key,
+                                      type: "classNames",
+                                      propertyName: propertyName,
+                                    });
+                                }}
+                                onMouseLeave={handleMouseLeave}
+                                onClick={() =>
+                                  mode === "click" &&
+                                  updateClassState({
+                                    elementKey: key,
+                                    type: "classNames",
+                                    propertyName: propertyName,
+                                  })
+                                }
+                              />
+                            )
+                          )}
+                        </div>
+                      )}
+                    </div>
+                  );
+                })}
+              </div>
+            )}
+          </ClassController>
         </div>
       </div>
     </StylePreviewerProvider>
@@ -112,3 +197,34 @@ const StylePreviewer = <T extends Record<string, ClassInfo>>({
 };
 
 export default StylePreviewer;
+
+const ClassItem = ({
+  name,
+  isSelected,
+  onMouseEnter,
+  onMouseLeave,
+  onClick,
+}: {
+  name: string;
+  isSelected: boolean;
+  onMouseEnter: () => void;
+  onMouseLeave: () => void;
+  onClick: () => void;
+}) => (
+  <div
+    onMouseEnter={onMouseEnter}
+    onMouseLeave={onMouseLeave}
+    onClick={onClick}
+    onKeyUp={(e) => e.key === "Enter" && onClick()}
+    style={{
+      padding: "8px",
+      border: "1px solid #ccc",
+      borderRadius: "4px",
+      marginBottom: "4px",
+      cursor: "pointer",
+      backgroundColor: isSelected ? "#f0f0f0" : "transparent",
+    }}
+  >
+    {name}
+  </div>
+);
