+++ src/components/StylePreview/ui/index.tsx
@@ -1,3 +1,3 @@
-export * from "./ClassSelectList";
+export * from "./ClassController";
 export * from "./HighlightedStyleElement";
 export * from "./ModeToggleController";
